+++ features/steps/currentcost_steps.py
@@ -7,9 +7,8 @@
     Functionnal test for SMEPI test cases.
 """
 
-#from behave import when, given, then  # pylint: disable-msg=E0611
 from __future__ import print_function
-from behave import when, then  # pylint: disable-msg=E0611
+from behave import when, then, given  # pylint: disable-msg=E0611
 import subprocess
 from subprocess import CalledProcessError
 from currentcost.utils import error_utils
@@ -22,7 +21,7 @@ BIN = "currentcost"
 VAR_NAME = "TEST_electric_meter"
 SITE_NAME = "TEST_liogen_home"
 ARGUMENT_TTY_PORT = "--tty-port"
-TTY_PORT = "/dev/currentcost"
+TTY_PORT = "tests/tty/currentcost"
 BAD_TTY_PORT = "/dev/currentcost9876"
 ARGUMENT_MQ_CREDENTIAL = "--rabbitMQ-credential"
 MQ_CREDENTIAL = "admin:password"
@@ -31,17 +30,42 @@ MQ_HOST = "localhost"
 LOG_FILE = "logs/currentcost.log"
 TTY_ERROR_MESSAGE = "None"
 CREDENTIALS = pika.PlainCredentials("admin", "password")
+SOCAT = "socat"
+SIMULATED_TTY_PORT = "PTY,link=%s" % TTY_PORT
+SIMULATED_TTY_PORT2 = "PTY,link=tests/tty/writer"
 CONNECTION = pika.BlockingConnection(
     pika.ConnectionParameters(host=MQ_HOST, credentials=CREDENTIALS))
+CURRENTCOST_MESSAGE = "<msg><src>CC128-v1.29</src><dsb>00786</dsb><time>00:31:36</time><tmpr>19.3</tmpr><sensor>0</sensor><id>00077</id><type>1</type><ch1><watts>00405</watts></ch1></msg>"
 
 
-def callback(channel, method, properties, body):
+
+def check_cc_unreachable(channel, method, properties, body):
     """
         Callback called when a new message is available.
     """
-    print("Callback => %s" % body)
+    print("check_cc_unreachable => %s" % body)
     assert body == error_utils.TTY_CONNECTION_PROBLEM % (
         VAR_NAME, SITE_NAME, BAD_TTY_PORT)
+    CONNECTION.close()
+
+
+def check_cc_unreachable2(channel, method, properties, body):
+    """
+        Callback called when a new message is available.
+    """
+    print("check_cc_unreachable => %s" % body)
+    assert body == error_utils.TTY_CONNECTION_PROBLEM % (
+        VAR_NAME, SITE_NAME, TTY_PORT)
+    CONNECTION.close()
+
+
+def check_cc_disconected(channel, method, properties, body):
+    """
+        Callback called when a new message is available.
+    """
+    print("check_cc_disconected => %s" % body)
+    assert body == error_utils.CURRENTCOST_TIMEOUT % (
+        VAR_NAME, SITE_NAME)
     CONNECTION.close()
 
 
@@ -52,13 +76,6 @@ def check_response_script(commands_response):
     for cmdr in commands_response:
         print("Response: %s" % cmdr)
         assert cmdr is not None
-
-
-def thread_subprocess():
-    """
-        Launch a subprocess in a non blocking way.
-    """
-    pass
 
 
 def launch_script(commands):
@@ -80,7 +97,7 @@ def launch_script(commands):
     return commands_response
 
 
-@when(u'we launch currentcost script without important argument')
+@when(u"we launch currentcost script without important argument")
 def when_launch_without_parameter(context):
     """
         Launch currentcost script without important argument.
@@ -95,7 +112,7 @@ def when_launch_without_parameter(context):
     context.commands_response = launch_script(commands)
 
 
-@then(u'we should see an error message on screen')
+@then(u"we should see an error message on screen")
 def error_script_without_parameter(context):
     """
         Except that subprocess raise an exception.
@@ -103,7 +120,7 @@ def error_script_without_parameter(context):
     check_response_script(context.commands_response)
 
 
-@when(u'we start currentcost with bad port without rabbitmq')
+@when(u"we start currentcost with bad port without rabbitmq")
 def when_launch_with_unreachable(context):
     """
         Launch currentcost script with wrong tty with -p active
@@ -118,7 +135,7 @@ def when_launch_with_unreachable(context):
     process.terminate()
 
 
-@when(u'we start currentcost with bad port and bad rabbitmq credential')
+@when(u"we start currentcost with bad port and bad rabbitmq credential")
 def launch_ccunreach_badrmq(context):
     """
         Launch unreachable currentcost with bad rabbitmq credential
@@ -134,7 +151,7 @@ def launch_ccunreach_badrmq(context):
     process.terminate()
 
 
-@then(u'we should see currentcost is unreachable in log')
+@then(u"we should see currentcost is unreachable in log")
 def detect_unreachability_log(context):
     """
         We should see currentcost unreachability in log file.
@@ -151,7 +168,7 @@ def detect_unreachability_log(context):
     assert last_log_file == error
 
 
-@then(u'we should see rabbitmq error in log')
+@then(u"we should see rabbitmq error in log")
 def detect_rabbitmqerror_log(context):
     """
         We should see currentcost unreachability in log file.
@@ -170,10 +187,10 @@ def detect_rabbitmqerror_log(context):
     assert last_log_file == error
 
 
-@when(u'we start currentcost with bad port with rabbitmq')
+@when(u"we start currentcost with bad port with rabbitmq")
 def launch_ccunreach_withoutrmq(context):
     """
-        Launch currentcost script with wrong tty with -p active
+        Launch currentcost script with wrong tty port.
     """
     commands = "%s %s %s %s %s %s %s" % (
         BIN, VAR_NAME, SITE_NAME, ARGUMENT_TTY_PORT, BAD_TTY_PORT,
@@ -183,15 +200,151 @@ def launch_ccunreach_withoutrmq(context):
     process.terminate()
 
 
-@then(u'we should receive a message saying that current cost is unreachable')
+@then(u"we should receive a message saying that current cost is unreachable")
 def receive_message_unreachable(context):
     """
-        Expect a message saying that currentcost is unreachable on 0MQ.
+        Expect a message saying that currentcost is unreachable on RabbitMQ.
     """
     channel = CONNECTION.channel()
     channel.queue_declare(queue=error_utils.ERROR)
     try:
-        channel.basic_consume(callback, queue=error_utils.ERROR, no_ack=True)
+        channel.basic_consume(
+            check_cc_unreachable,
+            queue=error_utils.ERROR,
+            no_ack=True)
         channel.start_consuming()
     except pika.exceptions.ConnectionClosed:
         pass
+
+
+@given(u"current cost does not send any message")
+def simulate_cc_disconnected(context):
+    """
+        Simulate USB port connection with socat.
+    """
+    commands = "%s %s %s" % (SOCAT, SIMULATED_TTY_PORT, SIMULATED_TTY_PORT2)
+    context.socat = subprocess.Popen(shlex.split(commands))
+
+
+@when(u"we launch currentcost script and reach the timeout limit")
+def cc_reach_timeout(context):
+    """
+        Launch currentcost script.
+    """
+    commands = "%s %s %s %s %s %s %s" % (
+        BIN, VAR_NAME, SITE_NAME, ARGUMENT_TTY_PORT, TTY_PORT,
+        ARGUMENT_MQ_CREDENTIAL, MQ_CREDENTIAL)
+    context.process = subprocess.Popen(shlex.split(commands))
+
+
+@then(u"we should get informed that current cost does not send messages")
+def rmq_no_messages(context):
+    """
+        Waited on RabbitMQ an error message saying that current cost
+        does not send any message.
+    """
+    CONNECTION = pika.BlockingConnection(
+        pika.ConnectionParameters(host=MQ_HOST, credentials=CREDENTIALS))
+    channel = CONNECTION.channel()
+    channel.queue_declare(queue=error_utils.ERROR)
+    try:
+        channel.basic_consume(
+            check_cc_disconected,
+            queue=error_utils.ERROR,
+            no_ack=True)
+        channel.start_consuming()
+    except pika.exceptions.ConnectionClosed:
+        pass
+
+    context.process.terminate()
+    context.socat.terminate()
+
+
+@then(u"we should see current cost does not send any message in log")
+def log_no_messages(context):
+    """
+        Waited to see in log that current cost does not send any messages.
+    """
+    log_file = open(LOG_FILE, "r")
+    lines = log_file.readlines()
+    log_file.close()
+    last_log_file = lines[-1].replace("\n", "").replace("\"", "")
+    last_log_file = " ".join(last_log_file.split(" ")[9:])
+    error = error_utils.CURRENTCOST_TIMEOUT % (
+        VAR_NAME, SITE_NAME)
+    print("Last line => %s" % last_log_file)
+    print("Expect => %s" % error)
+    assert last_log_file == error
+
+
+@given(u"current cost is connected and currentcost script is launched")
+def cc_launch_correctly(context):
+    """
+        Current cost tty port is created with socat and we launch currentcost
+        script.
+    """
+    commands = "%s %s %s" % (SOCAT, SIMULATED_TTY_PORT, SIMULATED_TTY_PORT2)
+    context.socat = subprocess.Popen(shlex.split(commands))
+
+    commands = "%s %s %s %s %s %s %s" % (
+        BIN, VAR_NAME, SITE_NAME, ARGUMENT_TTY_PORT, TTY_PORT,
+        ARGUMENT_MQ_CREDENTIAL, MQ_CREDENTIAL)
+    context.process = subprocess.Popen(shlex.split(commands))
+
+    sleep(3)
+
+    log_file = open(LOG_FILE, "r")
+    lines = log_file.readlines()
+    log_file.close()
+    last_log_file = lines[-1].replace("\n", "").replace("\"", "")
+    last_log_file = " ".join(last_log_file.split(" ")[9:])
+    error = error_utils.TTY_CONNECTION_SUCCESS % (
+        VAR_NAME, SITE_NAME, TTY_PORT)
+    print("Last line => %s" % last_log_file)
+    print("Expect => %s" % error)
+    assert last_log_file == error
+
+
+@when(u"we disconnect USB port")
+def usb_port_disconnection(context):
+    """
+        Simulation of an USB port disconnection killing socat process.
+    """
+    context.socat.terminate()
+
+
+@then(u"we should receive a message saying that current cost is disconnected")
+def receive_message_disconnected(context):
+    """
+        Expect a message saying that currentcost is unreachable on RabbitMQ.
+    """
+    CONNECTION = pika.BlockingConnection(
+        pika.ConnectionParameters(host=MQ_HOST, credentials=CREDENTIALS))
+    channel = CONNECTION.channel()
+    channel.queue_declare(queue=error_utils.ERROR)
+    try:
+        channel.basic_consume(
+            check_cc_unreachable2,
+            queue=error_utils.ERROR,
+            no_ack=True)
+        channel.start_consuming()
+    except pika.exceptions.ConnectionClosed:
+        pass
+    context.process.terminate()
+
+
+@then(u"we should see currentcost is disconnected in log")
+def detect_disconnected_log(context):
+    """
+        We should see currentcost unreachability in log file.
+    """
+    log_file = open(LOG_FILE, "r")
+    lines = log_file.readlines()
+    log_file.close()
+    last_log_file = lines[-1].replace("\n", "").replace("\"", "")
+    last_log_file = " ".join(last_log_file.split(" ")[9:])
+    error = error_utils.TTY_CONNECTION_PROBLEM % (
+        VAR_NAME, SITE_NAME, TTY_PORT)
+    print("Last line => %s" % last_log_file)
+    print("Expect => %s" % error)
+    assert last_log_file == error